+++ trunk/cpsycle/audio/src/duplicator2.c
@@ -20,13 +20,15 @@ static int parametervalue(psy_audio_Duplicator2*, int param);
 static void parameterrange(psy_audio_Duplicator2*, int param, int* minval, int* maxval);
 static int parameterlabel(psy_audio_Duplicator2*, char* txt, int param);
 static int parametername(psy_audio_Duplicator2*, char* txt, int param);
-static unsigned int numparameters(psy_audio_Duplicator2*);
+static uintptr_t numparameters(psy_audio_Duplicator2*);
 static unsigned int numparametercols(psy_audio_Duplicator2*);
 static void dispose(psy_audio_Duplicator2*);
 static uintptr_t numinputs(psy_audio_Duplicator2* self) { return 0; }
 static uintptr_t numoutputs(psy_audio_Duplicator2* self) { return 0; }
-static void loadspecific(psy_audio_Duplicator2*, struct psy_audio_SongFile*, unsigned int slot);
-static void savespecific(psy_audio_Duplicator2*, struct psy_audio_SongFile*, unsigned int slot);
+static void loadspecific(psy_audio_Duplicator2*, struct psy_audio_SongFile*,
+	uintptr_t slot);
+static void savespecific(psy_audio_Duplicator2*, struct psy_audio_SongFile*,
+	uintptr_t slot);
 
 static int transpose(int note, int offset);
 
@@ -244,7 +246,7 @@ int parametername(psy_audio_Duplicator2* self, char* txt, int param)
 	return 1;
 }
 
-unsigned int numparameters(psy_audio_Duplicator2* self)
+uintptr_t numparameters(psy_audio_Duplicator2* self)
 {
 	return 16 * 4;
 }
@@ -254,7 +256,8 @@ unsigned int numparametercols(psy_audio_Duplicator2* self)
 	return 4;	
 }
 
-void loadspecific(psy_audio_Duplicator2* self, struct psy_audio_SongFile* songfile, unsigned int slot)
+void loadspecific(psy_audio_Duplicator2* self, struct psy_audio_SongFile* songfile,
+	uintptr_t slot)
 {
 	uint32_t size;
